shipit_code_coverage: Move GitHub-related code into a separate module
